# v2/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EmailRequest": {
        "additionalProperties": false,
        "description": "Empty on purpose: there is no recipient. It comes from the stored email field.",
        "properties": {},
        "title": "EmailRequest",
        "type": "object"
      },
      "ErrorBody": {
        "description": "Every error: a fixed code and a fixed message. Never exception text or field values.",
        "properties": {
          "actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ActionView"
            },
            "title": "Actions",
            "type": "array"
          },
          "code": {
            "enum": [
              "invalid_request",
              "intake_not_found",
              "not_found",
              "method_not_allowed",
              "stale_turn",
              "action_not_available",
              "form_incomplete",
              "idempotency_key_required",
              "email_not_submitted",
              "email_no_address",
              "email_not_synthetic",
              "email_failed",
              "not_submitted",
              "resume_failed",
              "too_many_attempts",
              "internal_error"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "view": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TurnView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HelpView": {
        "description": "Fixed text for \"Talk to a person\" when no form is open (e.g. locked out).",
        "properties": {
          "info": {
            "items": {
              "type": "string"
            },
            "title": "Info",
            "type": "array"
          }
        },
        "required": [
          "info"
        ],
        "title": "HelpView",
        "type": "object"
      },
...
      "SessionView": {
        "description": "A new intake, or one opened on a new device with a resume code.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "view": {
            "$ref": "#/components/schemas/TurnView"
          }
        },
        "required": [
          "id",
          "token",
          "view"
        ],
        "title": "SessionView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Intake assistant (demo, synthetic data only)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/help/person": {
      "get": {
        "description": "What to do to talk to a person when no form is open (e.g. after a lockout).",
        "operationId": "person_help_help_person_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelpView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Person Help"
      }
    },
    "/intakes": {
      "post": {
        "operationId": "create_intake_intakes_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Intake"
      }
    },
    "/intakes/resume": {
      "post": {
        "description": "Open an intake on a new device. Failures get one fixed answer, whether or not the\ncode exists, and are rate-limited per code and per client.",
        "operationId": "resume_with_code_intakes_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Resume With Code"
      }
    },
    "/intakes/{intake_id}": {
      "get": {
        "operationId": "get_intake_intakes__intake_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Intake"
      }
    },
    "/intakes/{intake_id}/benefit-summary": {
      "post": {
        "operationId": "benefit_summary_intakes__intake_id__benefit_summary_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Benefit Summary"
      }
    },
    "/intakes/{intake_id}/email": {
      "post": {
        "description": "No recipient anywhere in this request: it is read from the stored email field.",
        "operationId": "send_email_intakes__intake_id__email_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EmailRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Send Email"
      }
    },
    "/intakes/{intake_id}/pause": {
      "post": {
        "operationId": "pause_intakes__intake_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Pause"
      }
    },
    "/intakes/{intake_id}/replies": {
      "post": {
        "operationId": "reply_intakes__intake_id__replies_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Reply"
      }
    },
    "/intakes/{intake_id}/resume": {
      "post": {
        "description": "PAUSED -> where the user was. NEEDS_HUMAN -> \"Continue on my own\".",
        "operationId": "resume_intakes__intake_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Resume"
      }
    },
    "/intakes/{intake_id}/resume-code": {
      "post": {
        "description": "The resume code on request, any time after starting (\"Take a break\" area).",
        "operationId": "resume_code_intakes__intake_id__resume_code_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-intake-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Intake-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeCodeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Resume Code"
      }
    },
...
    "/ui/text": {
      "get": {
        "description": "Fixed interface wording, so the frontend writes none of its own.",
        "operationId": "ui_text_ui_text_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UiText"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Ui Text"
      }
    }
  }
}